+++ NEWS.md
@@ -1,3 +1,9 @@
+# R62S3 2.0.0.9000
+
+* Big unexpected update! R62S3 has two new functions
+  * R62S4: Generate S4 generics and methods from R6 classes
+  * R62Fun: Generate functions only from R6 Classes
+
 # R62S3 1.4.2
 
 * Added better S3 generic/method compatibility by adding ... to method formals